Add files via upload
The purpose of this project is to use Long Short Term Memory, a recurrent neural network to predict the price of Ethereum, and using those predictions to create trade signals and backtest how the model and strategy performed. The ability to predict price movement is a problem that if solved can be very lucrative. We looked to expand on some of the data science techniques learned over the entirety of our program in order to predict the price of Ethereum.